shipping-service/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
requests
python-jose
PyJWT